🚀 Social media features added & GitHub Actions CI/CD pipeline fixes
✨ New Features:
- Social media handlers (posts, comments, likes, follows, profile)
- Performance benchmarks added
- Enhanced user models with social media fields

🐛 Bug Fixes:
- Fixed deprecated GitHub Actions (actions-rs/toolchain -> dtolnay/rust-toolchain)
- Extended JWT secrets to 64 characters for security
- Resolved all clippy lint warnings (redundant closures, unnecessary casts)
- Fixed method naming conflicts (from_str -> new_from_str)
- Updated error handling for auth service

🧪 Quality Improvements:
- All tests passing (44 tests)
- Clean clippy output
- Formatted code with cargo fmt
- Validated Docker builds
- Performance benchmarks ready

📋 Deployment Ready:
- Multi-environment deployment strategy
- Production-optimized Docker container
- Health checks implemented
- Quality gates enforced in CI/CD

+++ src/handlers/todo/list.rs
@@ -13,7 +13,7 @@ pub async fn handler(
     State(AppState { todo_service, .. }): State<AppState>,
     Extension(user): Extension<ContextUser>,
 ) -> impl IntoResponse {
-    match todo_service.list(user.user_id as i32).await {
+    match todo_service.list(user.user_id).await {
         Ok(result) => (
             StatusCode::OK,
             Json(JsonResponse::Success(
